requirements.txt
pytest ~= 8.3.2
pytest-playwright ~= 0.5.1
mimesis ~= 17.0.0
black ~= 24.8.0
flake8 ~= 7.1.1
pre-commit ~= 3.8.0